+++ src/error.ts
@@ -1,0 +1,5 @@
+class ShopifyError extends Error {}
+
+export {
+  ShopifyError
+};