+++ src/index.ts
@@ -1,1 +1,13 @@
-export { sayHello, sayGoodbye } from './world/hello-goodbye';
+import {Context} from './context';
+import {ShopifyError} from './error';
+
+const Shopify = {
+  Context,
+  ShopifyError,
+};
+
+export default Shopify;
+export {
+  Context,
+  ShopifyError,
+};